requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
addict=2.4.0=pypi_0
blas=1.0=mkl
brotli-python=1.0.9=py38h6a678d5_7
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.12.12=h06a4308_0
cachetools=5.3.2=pypi_0
causal-conv1d=1.0.0=pypi_0
certifi=2023.11.17=py38h06a4308_0
cffi=1.16.0=py38h5eee18b_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
cmake=3.28.1=pypi_0
contourpy=1.1.1=pypi_0
cryptography=41.0.7=py38hdda0065_0
cuda-cudart=11.7.99=0
cuda-cupti=11.7.101=0
cuda-libraries=11.7.1=0
cuda-nvrtc=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-runtime=11.7.1=0
cycler=0.12.1=pypi_0
einops=0.7.0=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py38h06a4308_0
fonttools=4.47.2=pypi_0
freetype=2.12.1=h4a9f257_0
fsspec=2023.12.2=pypi_0
future=0.18.3=pypi_0
giflib=5.2.1=h5eee18b_3
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py38heeb90bb_0
gnutls=3.6.15=he1e5248_0
google-auth=2.27.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
grpcio=1.60.0=pypi_0
h5py=3.10.0=pypi_0
huggingface-hub=0.20.3=pypi_0
idna=3.4=py38h06a4308_0
imageio=2.33.1=pypi_0
importlib-metadata=7.0.1=pypi_0
importlib-resources=6.1.1=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.2=py38h06a4308_0
jpeg=9e=h5eee18b_1
kiwisolver=1.4.5=pypi_0
lame=3.100=h7b6447c_0
lazy-loader=0.3=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcublas=11.10.3.66=0
libcufft=10.7.2.124=h4fbf590_0
libcufile=1.8.1.2=0
libcurand=10.3.4.107=0
libcusolver=11.4.0.1=0
libcusparse=11.7.4.91=0
libdeflate=1.17=h5eee18b_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libnpp=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_0
lit=17.0.6=pypi_0
lmdb=1.4.1=pypi_0
lz4-c=1.9.4=h6a678d5_0
mamba-ssm=1.0.1=pypi_0
markdown=3.5.2=pypi_0
markupsafe=2.1.3=py38h5eee18b_0
matplotlib=3.7.4=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py38h5eee18b_1
mkl_fft=1.3.8=py38h5eee18b_0
mkl_random=1.2.4=py38hdb19cb5_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py38h06a4308_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py38h06a4308_0
ninja=1.11.1.1=pypi_0
numpy=1.24.3=py38hf6e8229_1
numpy-base=1.24.3=py38h060ed82_1
oauthlib=3.2.2=pypi_0
opencv-python=4.9.0.80=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=3.0.12=h7f8727e_0
packaging=23.2=pypi_0
pillow=10.0.1=py38ha6cbd5a_0
pip=23.3.1=py38h06a4308_0
platformdirs=4.1.0=pypi_0
protobuf=4.25.2=pypi_0
pyasn1=0.5.1=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=23.2.0=py38h06a4308_0
pyparsing=3.1.1=pypi_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.18=h955ad1f_0
python-dateutil=2.8.2=pypi_0
pytorch=2.0.1=py3.8_cuda11.7_cudnn8.5.0_0
pytorch-cuda=11.7=h778d358_5
pytorch-mutex=1.0=cuda
pywavelets=1.4.1=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h5eee18b_0
regex=2023.12.25=pypi_0
requests=2.31.0=py38h06a4308_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
safetensors=0.4.2=pypi_0
scikit-image=0.21.0=pypi_0
scipy=1.10.1=pypi_0
setuptools=68.2.2=py38h06a4308_0
six=1.16.0=pypi_0
sqlite=3.41.2=h5eee18b_0
sympy=1.12=py38h06a4308_0
tb-nightly=2.14.0a20230808=pypi_0
tbb=2021.8.0=hdb19cb5_0
tensorboard-data-server=0.7.2=pypi_0
tifffile=2023.7.10=pypi_0
timm=0.9.12=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.15.1=pypi_0
tomli=2.0.1=pypi_0
torchaudio=2.0.2=py38_cu117
torchtriton=2.0.0=py38
torchvision=0.15.2=py38_cu117
tqdm=4.66.1=pypi_0
transformers=4.37.1=pypi_0
typing_extensions=4.9.0=py38h06a4308_1
urllib3=1.26.18=py38h06a4308_0
werkzeug=3.0.1=pypi_0
wheel=0.41.2=py38h06a4308_0
xz=5.4.5=h5eee18b_0
yapf=0.40.2=pypi_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
